getUserProfile in compoment Layout (not ProfilePage); Layout
