+++ src/pages/RegisterPage/RegisterPage.module.scss
@@ -56,7 +56,7 @@
           margin-block-end: 0;
           margin-left: 26.5px;
           margin-right: 26.5px;
-          margin-top: 36px;
+          // margin-top: 36px;
         }
         .p {
           font-weight: 400;
@@ -83,7 +83,7 @@
         }
   
         :global(.ant-form-item) {
-          margin-bottom: 10px;
+          margin-bottom: 7px;
         }
       }
